Refactor user invitation functionality by moving DTOs and service to Invitations namespace
