+++ aspnet-core/src/toyiyo.todo.Application/Invitations/Dto/UserInvitationDto.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using Abp.AutoMapper;
 using toyiyo.todo.Authorization.Users;
 
-namespace toyiyo.todo.Users
+namespace toyiyo.todo.Invitations.Dto
 {
     /// <summary>
     /// Data transfer object for user invitation requests.
