+++ aspnet-core/test/toyiyo.todo.Tests/Users/UserInvitationServiceTests.cs
@@ -5,6 +5,8 @@ using toyiyo.todo.Users;
 using Abp.UI;
 using toyiyo.todo.Authorization.Users;
 using toyiyo.todo.MultiTenancy;
+using toyiyo.todo.Invitations;
+using toyiyo.todo.Invitations.Dto;
 
 namespace toyiyo.todo.Tests.Users
 {
